Rename syntax to lezer-syntax
