+++ javascript/src/javascript.ts
@@ -1,6 +1,6 @@
 import parser from "lezer-javascript"
 import {StateExtension} from "../../state/src/"
-import {LezerSyntax} from "../../syntax/src/syntax"
+import {LezerSyntax} from "../../lezer-syntax/src/syntax"
 import {tokenTypes} from "../../highlight/src/highlight"
 import {syntaxIndentation, dontIndent, parens, braces, brackets, statement, compositeStatement} from "../../indent/src/indent"
